splitting routes constants and removing vite cache
