+++ package.json
@@ -13,8 +13,9 @@
     "build": "tsc && vite build --mode staging",
     "build:production": "tsc && vite build --mode production",
     "preview": "vite preview",
-    "test": "vitest run --coverage",
-    "test:update": "vitest run -u --coverage",
+    "test": "vitest run",
+    "test:update": "vitest run -u",
+    "test:coverage": "vitest run --coverage",
     "test:watch": "vitest",
     "test:ui": "vitest --ui",
     "cy:open": "cypress open",
