+++ src/context/LanguageProvider/__tests__/LanguageProvider.test.tsx
@@ -25,7 +25,7 @@ describe('LanguageProvider', () => {
     })
   })
   it('gir tilgang til react-intl translations', async () => {
-    const storeRef = await setupStore({}, true)
+    const storeRef = setupStore({}, true)
     render(
       <Provider store={storeRef}>
         <LanguageProvider>
@@ -41,7 +41,7 @@ describe('LanguageProvider', () => {
   })
 
   it('bruker locale fra cookie når den er tilgjengelig', async () => {
-    const storeRef = await setupStore({}, true)
+    const storeRef = setupStore({}, true)
     document.cookie = 'decorator-language=en'
     render(
       <Provider store={storeRef}>
