+++ src/pages/Step1/Step1Feil.tsx
@@ -3,8 +3,7 @@ import { useIntl } from 'react-intl'
 import { useNavigate } from 'react-router-dom'
 
 import { Card } from '@/components/common/Card'
-import { externalUrls } from '@/router'
-import { paths } from '@/router/routes'
+import { externalUrls, paths } from '@/router'
 import { useAppDispatch } from '@/state/hooks'
 import { userInputActions } from '@/state/userInput/userInputReducer'
