+++ src/router/routes.tsx
@@ -16,48 +16,12 @@ import { Step5 } from '@/pages/Step5'
 import { StepFeil } from '@/pages/StepFeil/'
 import { RouteErrorBoundary } from '@/router/RouteErrorBoundary'
 
+import { paths } from './constants'
 import {
   directAccessGuard,
   authenticationGuard,
   tpoMedlemskapAccessGuard,
 } from './loaders'
-
-export const BASE_PATH = '/pensjon/kalkulator'
-
-export const externalUrls = {
-  dinPensjon: 'https://nav.no/pensjon',
-  dinPensjonBeholdning: 'https://www.nav.no/pensjon/opptjening/nb/',
-  detaljertKalkulator: 'https://www.nav.no/pselv/simulering.jsf',
-  alderspensjonsregler: 'https://www.nav.no/alderspensjon#beregning',
-  afp: 'https://www.afp.no',
-  garantipensjon: 'https://www.nav.no/minstepensjon',
-  norskPensjon: 'https://norskpensjon.no/',
-  uinnloggetKalkulator:
-    'https://www.nav.no/pselv/simulering/forenkletsimulering.jsf',
-  personvernerklaering:
-    'https://www.nav.no/personvernerklaering#dine-rettigheter',
-  personvernerklaeringKontaktOss:
-    'https://www.nav.no/personvernerklaering#kontakt-nav',
-}
-
-export const paths = {
-  root: '/',
-  login: '/login',
-  henvisningUfoeretrygdGjenlevendepensjon:
-    '/henvisning-ufoeretrygd-gjenlevendepensjon',
-  henvisning1963: '/henvisning-1963',
-  start: '/start',
-  samtykke: '/samtykke',
-  utenlandsopphold: '/utenlandsopphold',
-  utenlandsoppholdFeil: '/henvisning-utland',
-  offentligTp: '/offentlig-tp',
-  afp: '/afp',
-  sivilstand: '/sivilstand',
-  uventetFeil: '/uventet-feil',
-  beregning: '/beregning',
-  forbehold: '/forbehold',
-  personopplysninger: '/personopplysninger',
-} as const
 
 export const routes: RouteObject[] = [
   {
@@ -165,3 +129,5 @@ export const routes: RouteObject[] = [
     ],
   },
 ]
+
+export default routes
